Make the project compilable when it's added as a git submodule
The previous gitBranch() function only search '$rootDir/.git/HEAD' file.
But, when the project is added as a git sub module, the '$rootDir/.git'
is a file, not a directory. The file stores the module git repository
directory. So the change is to help finding the submodule local git
repository.

Signed-off-by: Zhang Wei <weiz@lambdacloud.com>